pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vcan-test"
version = "0.1.0"
dependencies = [
    "python-can",
    "pytest"
]